third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/button-in-popover.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Buttons in the popover should be rendered and should not close the popover when clicked.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
[FAIL] Non-interactive content in the popover should not close the popover when clicked.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
Harness: the test ran to completion.
